pyproject.toml
[project]
name = "mcp-google-ads"
version = "0.1.0"
description = "Google Ads API integration for Model Context Protocol (MCP)"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Ernesto Cohnen", email = "ernesto@ixigo.com"}
]
keywords = ["mcp", "google ads", "seo", "sem", "claude","search analytics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
    "google-api-python-client>=2.163.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "mcp[cli]>=1.3.0",
]

[project.optional-dependencies]
web = [
    "uvicorn[standard]>=0.34.0",
    "google-auth-oauthlib>=1.2.0",
    "itsdangerous>=2.2.0",
    "starlette>=0.46.0",
    "httpx>=0.28.1",
]

[project.scripts]
google-ads-mcp-web = "web_server:run_web_server"

[project.urls]
"Homepage" = "https://github.com/cohnen/mcp-google-ads"
"Bug Tracker" = "https://github.com/cohnen/mcp-google-ads/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["mcp_google_ads"]
